Doom & DOS: give weapons properly by using IDKA and IDKFA
I think, this how it should work. Primary, this way player will not get unavailable weapons in Shareware. This is inspired by Doom Retro (thanks @bradharding!), but I want to keep gray 6 and 7 numbers on arms widget.

Not sure about real necessity of getting fist and pistol, but just in case…